+++ inst/include/dplyr/OrderVisitorImpl.h
@@ -55,6 +55,38 @@ namespace dplyr {
     private: 
         VECTOR vec ;    
     } ;
+    
+    template <bool ascending>
+    class StringOrderVisitor : public OrderVisitor {
+    public:
+        typedef OrderVectorVisitorImpl<INTSXP, ascending, IntegerVector> IntegerOrderVisitor ; 
+        
+        StringOrderVisitor( const CharacterVector& data_ ) : data(data_), visitor(orders(data)) {}
+        
+        inline bool equal(int i, int j) const {
+            return visitor.equal(i,j) ;    
+        }
+        
+        inline bool before(int i, int j) const {
+            return visitor.before(i,j) ;    
+        }
+        
+        inline SEXP get() {
+            return data ;    
+        }
+        
+    private:
+        
+        
+        inline IntegerVector orders( const CharacterVector strings ) {
+            Language call( "match", data, Language("sort", data) ) ;
+            return Rf_eval( call, R_GlobalEnv ) ;
+        }
+        
+        CharacterVector data ;
+        IntegerOrderVisitor visitor ;
+        
+    } ;
 
     inline OrderVisitor* order_visitor( SEXP vec, bool ascending ){
         if( ascending ){
@@ -62,7 +94,7 @@ namespace dplyr {
                 case INTSXP:  return new OrderVectorVisitorImpl<INTSXP , true, Vector<INTSXP > >( vec ) ;
                 case REALSXP: return new OrderVectorVisitorImpl<REALSXP, true, Vector<REALSXP> >( vec ) ;
                 case LGLSXP:  return new OrderVectorVisitorImpl<LGLSXP , true, Vector<LGLSXP > >( vec ) ;
-                case STRSXP:  return new OrderVectorVisitorImpl<STRSXP , true, Vector<STRSXP > >( vec ) ;
+                case STRSXP:  return new StringOrderVisitor<true>( vec ) ;
                 case CPLXSXP:  return new OrderVectorVisitorImpl<CPLXSXP , true, Vector<CPLXSXP > >( vec ) ;
                 default: break ;
             }
@@ -71,7 +103,7 @@ namespace dplyr {
                 case INTSXP:  return new OrderVectorVisitorImpl<INTSXP , false, Vector<INTSXP > >( vec ) ;
                 case REALSXP: return new OrderVectorVisitorImpl<REALSXP, false, Vector<REALSXP> >( vec ) ;
                 case LGLSXP:  return new OrderVectorVisitorImpl<LGLSXP , false, Vector<LGLSXP > >( vec ) ;
-                case STRSXP:  return new OrderVectorVisitorImpl<STRSXP , false, Vector<STRSXP > >( vec ) ;
+                case STRSXP:  return new StringOrderVisitor<false>( vec ) ;
                 case CPLXSXP:  return new OrderVectorVisitorImpl<CPLXSXP , false, Vector<CPLXSXP > >( vec ) ;
                 default: break ;
             }